Rework of base logic
