+++ .pomidory_settings.json
@@ -1,0 +1,1 @@
+{"pomodoroLength":25,"smallBreakLength":5,"bigBreakLength":20,"minimizeAfterStart":false,"theme":0}
