+++ view/save_view/save_view.go
@@ -5,7 +5,6 @@ import (
 	"fyne.io/fyne/v2/container"
 	"fyne.io/fyne/v2/layout"
 	"fyne.io/fyne/v2/widget"
-	"github.com/DK-92/pomidory/history"
 	"github.com/DK-92/pomidory/view"
 )
 
@@ -32,8 +31,8 @@ func CreateAndShowSaveView() {
 	app := view.GetAppInstance()
 	window = app.NewWindow(windowTitle)
 
-	totalHistory := history.GetInstance()
-	totalHistory.Save()
+	//totalHistory := history.GetInstance()
+	//totalHistory.Save()
 
 	vbox := container.New(
 		layout.NewVBoxLayout(),
